use hash builder for record encoder. Move JMapView to Attr Map
